move progress in its own package
